// pronti-frontend-main/front/src/app/modulos/finiquito/form/form-finiquito.enum.ts
export enum FormFiniquitoEnum {
  // nombre = 'nombre',
  idTrabajador = 'idTrabajador',
  fechaSalida = 'fechaSalida',
  motivoSalida = 'motivoSalida',
  documentoIdentidad = 'documentoIdentidad',
  fechaIngreso = 'fechaIngreso',
  profesion = 'profesion',
  sueldo = 'sueldo',
  fondoReservaIess = 'fondoReservaIess',
  formaPago = 'formaPago',

  credencial = 'credencial',
  descuentoCliente = 'descuentoCliente',
  stand = 'stand',
  chompa = 'chompa',
  clientesDesistidos = 'clientesDesistidos',
  revisor = 'revisor',
  aprobador = 'aprobador'
}

/*
* credencial, descuento cliente, stand, chompa, clientes desistidos, el revisor y para el aprobador
*/